src/grpc_in_the_snake_pit/tests/test_a.py
"""Tests on module 'a'"""

from grpc_in_the_snake_pit.a import A

class TestA:
    """
    Test Case for class 'A' in the 'a' module.
    """

    def test_get_value(self):
        """
        Test case for 'get_value' method.
        Assuming 'get_value' method returns the current value of the object.
        """
        obj = A(5)
        result = obj.get_value()
        assert result == 5

    def test_set_value(self):
        """
        Test case for 'set_value' method.
        Assuming 'set_value' method sets a new value for the object.
        """
        obj = A()
        obj.set_value(10)
        result = obj.get_value()
        assert result == 10

    def test_str(self):
        """
        Test case for '__str__' method.
        Assuming '__str__' method returns a string representation of the object.
        """
        obj = A(7)
        result = str(obj)
        assert result == "A(value=7)"
